bonjourmeal-codelab/full-sample/requirements.txt
Django==3.0.8
PyMySQL==0.10.0
google-apitools
google-auth-httplib2
django-extensions
google-businessmessages==1.0.0
stripe==2.54.0